refactor(commands): centralize option interaction logic
